pyproject.toml
[tool.poetry]
name = "syntech_test_task_short"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.2"
drf-spectacular = "^0.13.2"
celery = "^5.0.5"
django-filter = "^2.4.0"
redis = "^3.5.3"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
